chore: tidyup main.go
